[Feat] TabBarViewModel 생성, Deeplink로 앱 접속시 Comm Tab으로 이동, [Chore] MainTab의 home -> alert / alert -> comm으로 변경,  AppDelegate 파일분리, TabBar 디렉토리 분리
